requirements.txt
pandasai
langchain_anthropic
seaborn
python-dotenv
pandas
openpyxl
fastapi
uvicorn
python-multipart
aiofiles
